Remove Admin page and associated JSON configuration file
- Deleted the Admin.tsx component, which included various UI elements and state management for stages and sub-stages.
- Removed the P4-R2-12.json file that contained metadata and requirements for extracting seed and bootstrap logic from server routes, as this logic has been previously integrated into server initialization.
